==> ipynb/0_Overview of Result.ipynb <==
import pandas as pd
import warnings

rf_normal = pd.read_csv('../result/Random_forest.csv', index_col=0).drop('Time', axis=1)
xgb_normal = pd.read_csv('../result/XGBoost.csv', index_col=0).drop('Time', axis=1)
lgb_normal = pd.read_csv('../result/XGBoost.csv', index_col=0).drop('Time', axis=1)
keras_normal = pd.read_csv('../result/Keras.csv', index_col=0).drop('Time', axis=1)
keras_normal.Accuracy = round(keras_normal.Accuracy, 2)
t = pd.concat([rf_normal, xgb_normal, lgb_normal, keras_normal], axis=1).T
t.index = ['Random Forest', 'XGBoost', 'LightGBM', 'Keras']
k = pd.concat([pd.read_csv('../result/Feature_Selection_result.csv', index_col=0).T, pd.read_csv('../result/Feature_Selection_voting.csv', index_col=0)])
h = pd.concat([pd.read_csv('../result/Decomposition + HCD35.csv', index_col=0), pd.read_csv('../result/Decomposition + HCD35_voting.csv', index_col=0)])
n = round(pd.concat([pd.read_csv('../result/Feature Selection + Decomposition.csv', index_col=0)*100, pd.read_csv('../result/Feature Selection + Decomposition_voting.csv', index_col=0)]), 2)
gs = pd.read_csv('../result/voting_gs.csv', index_col=0)

### Simgle model / Voting

pd.concat([t, pd.read_csv('../result/voting.csv', index_col=0)]).sort_values('HCD35_Positive', ascending=False)

### Feature Selection & Decomposition / Voting

warnings.filterwarnings('ignore')
pd.concat([k, h, n], axis=1).sort_values('Feature_Selection_Positive', ascending=False)

### GridSearch / Voting

gs.sort_values('Feature Selection + Decomposition_Positive', ascending=False)
